README.md
Project title:Workforce Administration Solution (Dev)

Project Description:
Workforce Administration Solution is a software application or platform designed to streamline and automate various aspects of employee’s working on projects and Asset Assignment processes within an organization. It serves as a centralized system for managing employee data, number of projects an employee is working on, tracking employee performance, and keeping record for the assets which they are assigned to.